requirements.txt
django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.1.0
mysql-connector-python==8.1.0
requests==2.31.0
PyMySQL==1.1.0
djangorestframework-simplejwt==5.2.2
cryptography==41.0.3
python-dotenv==1.0.0
django_extensions==3.2.3